Split all non-Caja code off sanitizer.js to purifier.js for DCE.
